+++ srcs/parsing/check_walls.c
@@ -12,89 +12,25 @@
 
 #include "cub.h"
 
-
-//obsolete fct
-// static int	check_around_space(char **wall, int i, int j)
-// {
-// 	int	len;
-
-// 	if (wall[i][j + 1] != '1' && !is_space(wall[i][j + 1])
-// 		&& wall[i][j + 1] != 0 && wall[i][j + 1] != EOF)
-// 		return (1);
-// 	if (j > 0)
-// 		if (wall[i][j - 1] != '1' && !is_space(wall[i][j - 1]))
-// 			return (1);
-// 	len = ft_strlen(wall[i + 1]);
-// 	if (len > j)
-// 		if (wall[i + 1][j] != '1' && !is_space(wall[i + 1][j])
-// 			&& wall[i][j + 1] != 0 && wall[i][j + 1] != EOF)
-// 			return (1);
-// 	len = ft_strlen(wall[i - 1]);
-// 	if (len > j)
-// 		if (wall[i - 1][j] != '1' && !is_space(wall[i - 1][j])
-// 			&& wall[i][j + 1] != 0 && wall[i][j + 1] != EOF)
-// 			return (1);
-// 	return (0);
-// }
-
-// static int	check_middle_lines(char **wall, int max_size)
-// {
-// 	int	i;
-// 	int	j;
-// 	int	len;
-
-// 	i = 0;
-// 	while (wall[++i] && i < max_size)
-// 	{
-// 		if (check_left_wall(wall[i]) || check_right_wall(wall[i]))
-// 			return (1);
-// 		j = -1;
-// 		while (wall[i][++j])
-// 		{
-// 			if (wall[i][j] != '1')
-// 			{
-// 				if (is_space(wall[i][j]))
-// 				{
-// 					if (wall[i][j + 1] != '1' && !is_space(wall[i][j + 1]))
-// 						return (1);
-// 					len = ft_strlen(wall[i - 1]);
-// 					if (len >= j)
-// 					{
-// 						if (wall[i - 1][j] != '1' && !is_space(wall[i - 1][j]))
-// 							return (1);
-// 					}
-// 					else
-// 						return (1);
-// 					len = ft_strlen(wall[i + 1]);
-// 					if (len >= j)
-// 					{
-// 						if (wall[i + 1][j] != '1' && !is_space(wall[i + 1][j]))
-// 							return (1);
-// 					}
-// 					else
-// 						return (1);
-// 				}
-// 				else
-// 					if (((int)ft_strlen(wall[i + 1])) < j || ((int)ft_strlen(wall[i - 1])) < j)
-// 						return (1);
-// 			}
-// 		}
-// 	}
-// 	return (0);
-// }
-
-static int	is_player(char c)
+static int	adj_line(char *line, int j)
 {
-	if (c == 'N' || c == 'S' || c == 'E' || c == 'W')
+	int	len;
+
+	len = ft_strlen(line);
+	if (len >= j)
+	{
+		if (is_space(line[j]))
+			return (1);
+	}
+	else
 		return (1);
 	return (0);
 }
 
-static int	check_middle_lines(char **wall, int max_size)
+static int	middle_lines(char **wall, int max_size)
 {
 	int	i;
 	int	j;
-	int	len;
 
 	i = 0;
 	while (wall[++i] && i < max_size)
@@ -109,21 +45,7 @@ static int	check_middle_lines(char **wall, int max_size)
 				if (j > 0)
 				if (is_space(wall[i][j - 1]) || is_space(wall[i][j + 1]) || j == 0)
 					return (1);
-				len = ft_strlen(wall[i - 1]);
-				if (len >= j)
-				{
-					if (is_space(wall[i - 1][j]))
-						return (1);
-				}
-				else
-					return (1);
-				len = ft_strlen(wall[i + 1]);
-				if (len >= j)
-				{
-					if (is_space(wall[i + 1][j]))
-						return (1);
-				}
-				else
+				if (adj_line(wall[i - 1], j) || adj_line(wall[i + 1], j))
 					return (1);
 			}
 		}
@@ -131,59 +53,26 @@ static int	check_middle_lines(char **wall, int max_size)
 	return (0);
 }
 
-static int	check_last_line(char **wall, int i)
+static int	last_line(char **wall, int i)
 {
 	int	j;
-	// int	len;
 
 	j = -1;
-	// len = ft_strlen(wall[i - 1]);
 	while (wall[i][++j])
 	{
-		// if (is_space(wall[i][j]))
-		// {
-		// 	if (wall[i][j + 1] != '1' && !is_space(wall[i][j + 1])
-		// 		&& wall[i][j + 1] != 0 && wall[i][j + 1] != EOF)
-		// 		return (1);
-		// 	if (len >= j)
-		// 	{
-		// 		if (wall[i - 1][j] != '1' && !is_space(wall[i - 1][j]))
-		// 			return (1);
-		// 	}
-		// 	else
-		// 		return (1);
-		// }
-		// else if (wall[i][j] != '1')
-		// 	return (1);
 		if (wall[i][j] != '1' && is_space(wall[i][j]))
 		 	return (1);
 	}
 	return (0);
 }
 
-static int	check_first_line(char **wall)
+static int	first_line(char **wall)
 {
 	int	x;
-	// int	len;
 
 	x = -1;
-	// len = ft_strlen(wall[1]);
 	while (wall[0][++x])
 	{
-		// if (is_space(wall[0][x]))
-		// {
-		// 	if (wall[0][x + 1] != '1' && !is_space(wall[0][x + 1]))
-		// 		return (1);
-		// 	if (len >= x)
-		// 	{
-		// 		if (wall[1][x] != '1' && !is_space(wall[1][x]))
-		// 			return (1);
-		// 	}
-		// 	else
-		// 		return (1);
-		// }
-		// else if (wall[0][x] != '1')
-		// 	return (1);
 		if (wall[0][x] != '1' && !is_space(wall[0][x]))
 			return (1);
 	}
@@ -193,10 +82,10 @@ static int	check_first_line(char **wall)
 int	check_walls(t_file *file)
 {
 	print_map(file->scene);
-	if (check_first_line(file->scene)
-		|| check_last_line(file->scene, ft_lstsize(file->map) - 1))
+	if (first_line(file->scene)
+		|| last_line(file->scene, ft_lstsize(file->map) - 1))
 		return (write(2, "Error\nMap not walled in\n", 25), 1);
-	if (check_middle_lines(file->scene, ft_lstsize(file->map) - 1))
+	if (middle_lines(file->scene, ft_lstsize(file->map) - 1))
 		return (write(2, "Error\nMap not walled in\n", 25), 1);
 	return (0);
 }
