+++ srcs/parsing/convert_list_to_array.c
@@ -86,37 +86,6 @@ static int	check_for_newline(t_list *map)
 	return (0);
 }
 
-void	free_scene(t_file *file, int i)
-{
-	while (file->scene[--i])
-		free(file->scene[i]);
-	free(file->scene);
-}
-
-int	trim_end_spaces(t_file *file)
-{
-	int		y;
-	int		x;
-	int		len;
-	char	*tmp;
-
-	y = -1;
-	while (file->scene[++y])
-	{
-		x = 0;
-		len = ft_strlen(file->scene[y]);
-		while (is_space(file->scene[y][--len]))
-			x++;
-		tmp = ft_strdup(file->scene[y]);
-		free(file->scene[y]);
-		file->scene[y] = ft_substr(tmp, 0, ft_strlen(tmp) - x);
-		free(tmp);
-		if (!file->scene[y])
-			return (write(2, "Error\nMalloc failed.\n", 21), 1);
-	}
-	return (0);
-}
-
 int	convert_list_to_array(t_file *file)
 {
 	int		i;
@@ -137,11 +106,7 @@ int	convert_list_to_array(t_file *file)
 		i++;
 	}
 	file->scene[i] = 0;
-	printf("BEFORE\n");
-	print_map(file->scene);
 	if (trim_end_spaces(file))
 		return (1);
-	printf("AFTER\n");
-	print_map(file->scene);
 	return (0);
 }
